// src/components/Typography/TypographyVariants.d.ts
export declare const TypographyVariants: (props?: ({
    variant?: "title" | "display" | "body" | "caption" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "displayXl" | "subtitle" | "bodyLg" | "bodySm" | "overline" | "heading1" | "heading2" | "heading3" | "heading4" | "heading5" | "heading6" | "bodyLarge" | "bodySmallAlias" | null | undefined;
    weight?: "bold" | "light" | "normal" | "medium" | "semibold" | "extrabold" | "black" | null | undefined;
    color?: "default" | "primary" | "muted" | "success" | "danger" | "white" | null | undefined;
    align?: "end" | "center" | "start" | null | undefined;
    truncate?: boolean | null | undefined;
    lineClamp?: 1 | 2 | 3 | 4 | 5 | null | undefined;
} & import("class-variance-authority/types").ClassProp) | undefined) => string;